+++ README.md
@@ -17,6 +17,8 @@ For running this project locally in your system, please go through steps below:
 
     npm run start - for production mode
 ```
+    - Websocket server runs with default port on ui or it's can be filled in the input with value :8080
+
 2) Run backend:
 
 - Open new terminal and use next commands:
@@ -29,3 +31,4 @@ For running this project locally in your system, please go through steps below:
 
     npm run dev
 ```
+
