+++ src/controller/currentPosition.js
@@ -1,0 +1,6 @@
+import robot from 'robotjs';
+
+export const getCurrentPositions = () => {  
+    const { x, y }  = robot.getMousePos();
+    return `mouse_position ${x},${y}`;
+}
