+++ src/controller/updateMousePosition.js
@@ -1,0 +1,5 @@
+import robot from 'robotjs';
+
+export const updateMousePosition = (x, y) => {      
+    robot.moveMouse(x, y);   
+}
